修改完成

+++ ChildManager/UI/cepingshi/cp_yysc1_panel.cs
@@ -248,9 +248,9 @@ namespace ChildManager.UI.cepingshi
         {
             Cursor.Current = Cursors.WaitCursor;
             _cdiobj = cdibll.Get(id, _type);
-            _cdi1obj = cdi1bll.Get(_cpwomeninfo.cd_id,  _type, _cdiobj?.ID ?? 0);
-            _ddstobj = ddstbll.GetByType(_cpwomeninfo.cd_id, _type, _cdiobj?.ID ?? 0);
-            _zqobj = ppvtbll.GetByType(_cpwomeninfo.cd_id, _type, _cdiobj?.ID ?? 0);
+            _cdi1obj = cdi1bll.Get(_cpwomeninfo.cd_id,  _type,(int)(_cdiobj?.ID ?? 0));
+            _ddstobj = ddstbll.GetByType(_cpwomeninfo.cd_id, _type,(int)(_cdiobj?.ID ?? 0));
+            _zqobj = ppvtbll.GetByType(_cpwomeninfo.cd_id, _type, (int)(_cdiobj?.ID ?? 0));
             if (_cdiobj != null)
             {
                 CommonHelper.setForm(_cdiobj, panel1.Controls);
@@ -296,7 +296,7 @@ namespace ChildManager.UI.cepingshi
                     Cursor.Current = Cursors.WaitCursor;
                     try
                     {
-                        if (cdibll.Delete(_cdiobj.ID) && cdi1bll.DeleteByExternalid(_cdiobj.ID) && ddstbll.DeleteByExternalid(_cdiobj.ID) && ppvtbll.DeleteByExternalid(_cdiobj.ID))
+                        if (cdibll.Delete(_cdiobj.ID) && cdi1bll.DeleteByExternalid((int)_cdiobj.ID) && ddstbll.DeleteByExternalid((int)_cdiobj.ID) && ppvtbll.DeleteByExternalid((int)_cdiobj.ID))
                         {
                             MessageBox.Show("删除成功!", "软件提示");
                             RefreshRecordList();
