+++ YCF.Model/mb_zd.cs
@@ -15,7 +15,7 @@ namespace YCF.Model
     public partial class MB_ZD
     {
         public decimal ID { get; set; }
-        public int YFKS { get; set; }
+        public decimal YFKS { get; set; }
         public decimal YFJS { get; set; }
         public string WYZD { get; set; }
         public string ZJZD { get; set; }
